+++ src/ts/Browser.ts
@@ -114,6 +114,8 @@ const setIosDeviceInfo = () => {
 
 if(window.parent !== window) {
     const abGroupElement = <HTMLInputElement>window.parent.document.getElementById('abGroup');
+    const campaignIdElement = <HTMLInputElement>window.parent.document.getElementById('campaignId');
+    const countryElement = <HTMLInputElement>window.parent.document.getElementById('country');
     const platformElement = <HTMLInputElement>window.parent.document.getElementById('platform');
     const gameIdElement = <HTMLInputElement>window.parent.document.getElementById('gameId');
     const testModeElement = <HTMLInputElement>window.parent.document.getElementById('testMode');
@@ -124,6 +126,8 @@ if(window.parent !== window) {
         event.preventDefault();
 
         abGroupElement.disabled = true;
+        campaignIdElement.disabled = true;
+        countryElement.disabled = true;
         platformElement.disabled = true;
         gameIdElement.disabled = true;
         testModeElement.disabled = true;
@@ -137,6 +141,18 @@ if(window.parent !== window) {
         if(abGroupElement.value.length) {
             publicStorage.test.abGroup = {
                 value: parseInt(abGroupElement.value, 10),
+                ts: Date.now()
+            };
+        }
+        if(campaignIdElement.value.length) {
+            publicStorage.test.campaignId = {
+                value: campaignIdElement.value,
+                ts: Date.now()
+            };
+        }
+        if(countryElement.value.length) {
+            publicStorage.test.country = {
+                value: countryElement.value,
                 ts: Date.now()
             };
         }